+++ app/models/project.rb
@@ -113,7 +113,7 @@ class Project < ActiveRecord::Base
 
   def users_to_email
     result = (self.users + [self.user] + self.sites.collect{|s| s.users}.flatten).uniq
-    result = result.select{ |u| u.email_on?(:send_email) }
+    result = result.select{ |u| u.emails_enabled? }
   end
 
   # Returns "fake" constructed variables like 'site' and 'sheet_date'