+++ config/initializers/contour.rb
@@ -160,7 +160,7 @@ Contour.setup do |config|
 
   # An array of hashes that specify additional fields to add to the sign up form
   # An example might be [ { attribute: 'first_name', type: 'text_field' }, { attribute: 'last_name', type: 'text_field' } ]
-  config.sign_up_fields = [ { attribute: 'first_name', type: 'text_field' }, { attribute: 'last_name', type: 'text_field' } ]
+  config.sign_up_fields = [ { attribute: 'first_name', type: 'text_field' }, { attribute: 'last_name', type: 'text_field' }, { attribute: 'emails_enabled', type: 'check_box' } ]
 
   # An array of text fields used to trick spam bots using the honeypot approach. These text fields will not be displayed to the user.
   # An example might be [ :url, :address, :contact, :comment ]