+++ db/migrate/20150825174519_add_emails_enabled_to_users.rb
@@ -1,0 +1,5 @@
+class AddEmailsEnabledToUsers < ActiveRecord::Migration
+  def change
+    add_column :users, :emails_enabled, :boolean, null: false, default: false
+  end
+end